+++ Samples/SampleDiningPhilosophers/Fork.swift
