+++ Samples/SampleDiningPhilosophers/Philosopher.swift
