+++ Samples/SampleDiningPhilosophers/dining-philosopher-fsm.graffle
